[lightsoff] Reorganize order of widget realization; fixes LO on Clutter HEAD
I've been chasing after this for two days now, at least, unable to get Lights Off to run.
It turns out that you need to realize the window before the child stage, and both of
these things need to happen before attempting to realize any child ClutterActors.

This means I don't know the size of the window when I'm showing it (yet), so (at least
with xmonad), the window is positioned oddly.