[lightsoff] Ensure level is never < 1 (causes infinite loop)
For some reason, it seems that if the machine (or gconfd, or something)
isn't restarted after installing schemas, they don't take effect.

This was causing us to successfully load a score of 0, locking up the
load_level function.

I believe that the libgames-support conf stuff has defaults in itself, so
once we switch to that, that will also solve this problem. Still, there
must be some form of solution to the GConf schema loading problem too...