Fixed License section of Help
Fixes #615794

There is a strange bug in yelp that means that
<citetitle> cannot occur within <ulink>: otherwise
the ulink's url parameter is rendered. This commit works
around this bug by placing the <citetitle> outside the
<ulink>.

Manual testing of the original has been undertaken -
the text renders as desired. Accuracy of PO files has
NOT yet been tested.