Properly include the icons directory from Makefile.am
I'm not entirely sure how anyone could have made a release without this
change but there must be someway of doing so. This commit just formally
adds the icons directory to the distcheck dependency chain so that the
configure.ac references to the icon Makefile outputs show up in the
final tarball directory structure.

Not incrementing the release for this.
(cherry picked from commit 1b2fa577ad58dcbe29acbdb32df1208af0d5b815)